# packages/core/mcp/mcp-server.ts
// Removed #!/usr/bin/env node - this is now a module
import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
import { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
import { isInitializeRequest } from "@modelcontextprotocol/sdk/types.js";
import { RequestSource, ToolResult } from "@superglue/shared";
import { randomUUID } from "crypto";
import { Request, Response } from "express";
import { z } from "zod";
import { checkToolExecutionPermission } from "../api/ee/index.js";
import { validateToken } from "../auth/auth.js";
import { logMessage } from "../utils/logs.js";
import { sessionId, telemetryClient } from "../utils/telemetry.js";
import { validateWorkflowExecutionArgs } from "./mcp-server-utils.js";

// MCP Tool Input Schemas (tool-centric)
export const FindRelevantToolsInputSchema = z.object({
  searchTerms: z.string().describe("The natural language search query to find relevant tools"),
});

export const ExecuteToolInputSchema = z.object({
  id: z.string().describe("The ID of the tool to execute"),
  payload: z
    .record(z.string(), z.unknown())
    .optional()
    .describe("JSON payload to pass to the tool"),
});

export const AuthenticateInputSchema = z.object({});

// REST API client for MCP operations
class McpRestClient {
  private apiEndpoint: string;
  private apiKey: string;

  constructor(apiKey: string) {
    this.apiKey = apiKey;
    // Use API_ENDPOINT env var, fallback to localhost
    const port = process.env.API_PORT || "3002";
    this.apiEndpoint = process.env.API_ENDPOINT || `http://localhost:${port}`;
  }

  private async request<T>(method: string, path: string, body?: any): Promise<T> {
    const url = `${this.apiEndpoint.replace(/\/$/, "")}/v1${path}`;
    const headers: Record<string, string> = {
      Authorization: `Bearer ${this.apiKey}`,
    };

    if (body && method !== "GET") {
      headers["Content-Type"] = "application/json";
    }

    const response = await fetch(url, {
      method,
      headers,
      body: body ? JSON.stringify(body) : undefined,
    });

    if (!response.ok) {
      const errorText = await response.text();
      let errorMessage = `HTTP ${response.status}: ${errorText}`;
      try {
        const errorJson = JSON.parse(errorText);
        errorMessage = errorJson.error || errorJson.message || errorMessage;
      } catch {
        // ignore parse error
      }
      throw new Error(errorMessage);
    }

    return response.json();
  }

  async runTool(args: { id: string; payload?: Record<string, unknown> }): Promise<any> {
    const result = await this.request<any>("POST", `/tools/${encodeURIComponent(args.id)}/run`, {
      inputs: args.payload,
      options: { requestSource: RequestSource.MCP },
    });
    return {
      success: result.status === "success",
      data: result.data,
      error: result.error,
    };
  }

  async listTools(): Promise<any[]> {
    const result = await this.request<{ data: any[] }>("GET", "/tools?limit=1000");
    return result.data || [];
  }

  async generatePortalLink(): Promise<{ success: boolean; portalUrl?: string; error?: string }> {
    try {
      // Generate a portal link for the end user to authenticate with all systems
      const result = await this.request<{
        success: boolean;
        data: { portalUrl: string; token: string; expiresAt: string };
      }>("POST", `/authenticate`, {});
      return {
        success: true,
        portalUrl: result.data.portalUrl,
      };
    } catch (error: any) {
      return {
        success: false,
        error: error.message,
      };
    }
  }
}

// Update execute functions with validation
export const toolDefinitions: Record<string, any> = {
  superglue_execute_tool: {
    description: `
    <use_case>
      Executes an existing superglue tool by its ID.
    </use_case>

    <important_notes>
      - This tool is for running existing, saved superglue tools.
      - The superglue tool ID must exist, otherwise the tool will fail to execute.
      - The tool will return the result of the execution.
    </important_notes>
    `,
    inputSchema: ExecuteToolInputSchema,
    execute: async (args: any & { client: McpRestClient; orgId: string }, request) => {
      const validationErrors = validateWorkflowExecutionArgs(args);

      if (validationErrors.length > 0) {
        return {
          success: false,
          error: validationErrors.join("\n"),
        };
      }

      try {
        const result: ToolResult & { data?: any } = await args.client.runTool({
          id: args.id,
          payload: args.payload,
        });

        if (!result.success) {
          return {
            success: false,
            error: result.error || "Unknown error",
          };
        }

        const data = result.data;
        if (data === undefined || data === null) {
          return {
            success: true,
            data: null,
          };
        }

        const dataStr = JSON.stringify(data);
        const limit = 20000;

        if (dataStr.length <= limit) {
          return {
            success: true,
            data,
          };
        }

        return {
          success: true,
          data: `[TRUNCATED: Result exceeded ${limit} characters (original size: ${dataStr.length} chars). Showing first ${limit} characters]\n\n${dataStr.slice(0, limit)}`,
        };
      } catch (error: any) {
        return {
          success: false,
          error: error.message,
          suggestion: "Check that the tool ID exists and all required credentials are provided",
        };
      }
    },
  },
  superglue_find_relevant_tools: {
    description: `
    <use_case>
      Finds relevant superglue tools based on natural language search terms.
    </use_case>

    <important_notes>
      - This tool is for finding relevant superglue tools based on natural language search terms.
      - The tool will return a list of superglue tool IDs that match the search terms.
      - The tool will return a list of all superglue tool IDs if no search terms are provided, or if there are no matches for the search terms.
      - Use '*' as a wildcard to find all tools.
    </important_notes>
    `,
    inputSchema: FindRelevantToolsInputSchema,
    execute: async (args: any & { client: McpRestClient; orgId: string }, request) => {
      try {
        // List all tools via REST API - already filtered by permission at the API level
        const tools = await args.client.listTools();
        const activeTools = tools?.filter((t: any) => !t.archived);

        return {
          success: true,
          tools: activeTools.map((t: any) => ({
            id: t.id,
            instruction: t.instruction,
            inputSchema: t.inputSchema?.properties?.payload,
            responseSchema: t.outputSchema,
            steps: (t.steps || []).map((s: any) => ({
              systemId: s.systemId,
              stepId: s.id,
            })),
          })),
        };
      } catch (error: any) {
        return {
          success: false,
          error: error.message,
          suggestion: "Check that the query is valid",
        };
      }
    },
  },
  superglue_authenticate: {
    description: `
    <use_case>
      Generates an authentication portal link for an end user to connect their accounts.
      Use this when a tool execution fails because the end user hasn't authenticated with required systems.
    </use_case>

    <important_notes>
      - This tool generates a secure, time-limited link to the authentication portal.
      - The link should be shared with the end user so they can connect their accounts.
      - The portal allows users to authenticate with all available systems that require credentials.
      - After the user authenticates, they will be able to use tools that require those systems.
      - The API key must be linked to an end user for this to work.
    </important_notes>
    `,
    inputSchema: AuthenticateInputSchema,
    execute: async (args: { client: McpRestClient; orgId: string }, request: any) => {
      try {
        const result = await args.client.generatePortalLink();

        if (!result.success) {
          return {
            success: false,
            error: result.error,
            suggestion: "Ensure the API key is linked to an end user",
          };
        }

        return {
          success: true,
          portalUrl: result.portalUrl,
          message: `Please share this link with the user to authenticate: ${result.portalUrl}`,
        };
      } catch (error: any) {
        return {
          success: false,
          error: error.message,
          suggestion: "Ensure the API key is linked to an end user",
        };
      }
    },
  },
};

interface McpAuthContext {
  orgId: string;
  isRestricted?: boolean;
  allowedTools?: string[];
}

export const createMcpServer = async (apiKey: string) => {
  const mcpServer = new McpServer(
    {
      name: "superglue",
      version: "0.1.0",
    },
    {
      capabilities: {
        logging: {},
        tools: {},
      },
    },
  );

  const client = new McpRestClient(apiKey);

  const getAuthContext = async (): Promise<McpAuthContext> => {
    const authResult = await validateToken(apiKey);
    return {
      orgId: authResult.orgId,
      isRestricted: authResult.isRestricted,
      allowedTools: authResult.allowedTools,
    };
  };

  // Register tools individually for proper type inference
  mcpServer.registerTool(
    "execute_tool",
    {
      description: toolDefinitions.superglue_execute_tool.description,
      inputSchema: ExecuteToolInputSchema,
    },
    async (args, extra) => {
      // Re-validate to get fresh permissions on each call
      const authContext = await getAuthContext();

      // EE: Check if this API key is authorized to execute this tool
      const permCheck = checkToolExecutionPermission(authContext, args.id);
      if (!permCheck.allowed) {
        logMessage("warn", `MCP: API key not authorized to execute tool ${args.id}`, {
          orgId: authContext.orgId,
        });
        return {
          content: [
            {
              type: "text" as const,
              text: JSON.stringify({
                success: false,
                error: permCheck.error || "This API key is not authorized to execute this tool",
              }),
            },
          ],
        };
      }

      const result = await toolDefinitions.superglue_execute_tool.execute(
        { ...args, client, orgId: authContext.orgId },
        extra,
      );
      logMessage("debug", "superglue_execute_tool executed via MCP", { orgId: authContext.orgId });
      telemetryClient?.capture({
        distinctId: authContext.orgId || sessionId,
        event: "mcp_superglue_execute_tool",
        properties: {
          toolName: "superglue_execute_tool",
          orgId: authContext.orgId,
        },
      });
      return {
        content: [
          {
            type: "text" as const,
            text: JSON.stringify(result),
          },
        ],
      };
    },
  );

  mcpServer.registerTool(
    "find_relevant_tools",
    {
      description: toolDefinitions.superglue_find_relevant_tools.description,
      inputSchema: FindRelevantToolsInputSchema,
    },
    async (args, extra) => {
      // Re-validate to get fresh permissions on each call
      const authContext = await getAuthContext();

      const result = await toolDefinitions.superglue_find_relevant_tools.execute(
        { ...args, client, orgId: authContext.orgId },
        extra,
      );

      logMessage("debug", "superglue_find_relevant_tools executed via MCP", {
        orgId: authContext.orgId,
      });
      telemetryClient?.capture({
        distinctId: authContext.orgId || sessionId,
        event: "mcp_superglue_find_relevant_tools",
        properties: {
          toolName: "superglue_find_relevant_tools",
          orgId: authContext.orgId,
        },
      });
      return {
        content: [
          {
            type: "text" as const,
            text: JSON.stringify(result),
          },
        ],
      };
    },
  );

  // EE: Register authenticate tool for multi-tenancy
  mcpServer.registerTool(
    "authenticate",
    {
      description: toolDefinitions.superglue_authenticate.description,
      inputSchema: AuthenticateInputSchema,
    },
    async (args, extra) => {
      const authContext = await getAuthContext();

      const result = await toolDefinitions.superglue_authenticate.execute(
        { ...args, client, orgId: authContext.orgId },
        extra,
      );

      logMessage("debug", "superglue_authenticate executed via MCP", {
        orgId: authContext.orgId,
      });
      telemetryClient?.capture({
        distinctId: authContext.orgId || sessionId,
        event: "mcp_superglue_authenticate",
        properties: {
          toolName: "superglue_authenticate",
          orgId: authContext.orgId,
        },
      });
      return {
        content: [
          {
            type: "text" as const,
            text: JSON.stringify(result),
          },
        ],
      };
    },
  );

  return mcpServer;
};

export const transports: { [sessionId: string]: StreamableHTTPServerTransport } = {};

export const mcpHandler = async (req: Request, res: Response) => {
  // Check for existing session ID
  const sessionId = req.headers["mcp-session-id"] as string | undefined;
  let transport: StreamableHTTPServerTransport;

  if (sessionId && transports[sessionId]) {
    // Reuse existing transport
    transport = transports[sessionId];
  } else if (!sessionId && isInitializeRequest(req.body)) {
    // New initialization request
    transport = new StreamableHTTPServerTransport({
      sessionIdGenerator: () => randomUUID(),
      onsessioninitialized: (sessionId) => {
        // Store the transport by session ID
        transports[sessionId] = transport;
      },
    });

    // Clean up transport when closed
    transport.onclose = () => {
      if (transport.sessionId) {
        delete transports[transport.sessionId];
      }
    };
    const token = (req as any).authInfo.token;
    const server = await createMcpServer(token);

    // Connect to the MCP server
    await server.connect(transport);
  } else {
    // Invalid request
    res.status(400).json({
      jsonrpc: "2.0",
      error: {
        code: -32000,
        message: "Bad Request: No valid session ID provided",
      },
      id: null,
    });
    return;
  }

  // Handle the request
  await transport.handleRequest(req, res, req.body);
};